requirements.txt
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
filelock==3.8.0
Flask==2.2.2
Flask-MySQL==1.5.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lshw==0.0.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
platformdirs==2.5.2
psutil==5.9.1
pyFirmata==1.1.0
PyMySQL==1.0.2
pynput==1.7.6
pyserial==3.5
six==1.16.0
urllib3==1.26.12
virtualenv==20.16.5
Werkzeug==2.2.2
